requirements.txt
discord.py
jira
supabase
python-dotenv
